refactor(api): Factor out core match creation logic for re-use
