+++ src/server/api.ts
@@ -11,12 +11,10 @@ import Router from 'koa-router';
 import koaBody from 'koa-body';
 import { generate as shortid } from 'shortid';
 import cors from '@koa/cors';
-
-import { InitializeGame } from '../core/initialize';
 import { Auth } from './auth';
 import * as StorageAPI from './db/base';
 import { Server, LobbyAPI, Game } from '../types';
-import { createMetadata } from './util';
+import { createMatch } from './util';
 
 /**
  * Creates a new match.
@@ -29,29 +27,23 @@ import { createMetadata } from './util';
  * @param {object } lobbyConfig - Configuration options for the lobby.
  * @param {boolean} unlisted - Whether the match should be excluded from public listing.
  */
-export const CreateMatch = async ({
+const CreateMatch = async ({
+  ctx,
   db,
-  game,
-  numPlayers,
-  setupData,
-  uuid,
-  unlisted,
+  ...opts
 }: {
   db: StorageAPI.Sync | StorageAPI.Async;
-  game: Game;
-  numPlayers: number;
-  setupData: any;
-  uuid: () => string;
-  unlisted: boolean;
-}) => {
-  if (!numPlayers || typeof numPlayers !== 'number') numPlayers = 2;
+  ctx: Koa.BaseContext;
+} & Parameters<typeof createMatch>[0]) => {
+  const match = createMatch(opts);
 
-  const metadata = createMetadata({ game, numPlayers, setupData, unlisted });
-  const matchID = uuid();
-  const initialState = InitializeGame({ game, numPlayers, setupData });
+  if ('setupDataError' in match) {
+    ctx.throw(400, match.setupDataError);
+    return;
+  }
 
-  await db.createMatch(matchID, { metadata, initialState });
-
+  const { matchID, ...data } = match;
+  await db.createMatch(matchID, data);
   return matchID;
 };
 
@@ -123,11 +115,8 @@ export const createRouter = ({
     const game = games.find(g => g.name === gameName);
     if (!game) ctx.throw(404, 'Game ' + gameName + ' not found');
 
-    const setupDataError =
-      game.validateSetupData && game.validateSetupData(setupData, numPlayers);
-    if (setupDataError !== undefined) ctx.throw(400, setupDataError);
-
     const matchID = await CreateMatch({
+      ctx,
       db,
       game,
       numPlayers,
@@ -365,6 +354,7 @@ export const createRouter = ({
 
     const game = games.find(g => g.name === gameName);
     const nextMatchID = await CreateMatch({
+      ctx,
       db,
       game,
       numPlayers,
